requirements.txt
Flask~=2.0.1
Flask-SQLAlchemy
Ariadne
Flask-Cors
python-dotenv